# perf: use custom sqlite library on unix