--- a/.gitmodules
+++ b/.gitmodules
@@ -1,0 +1,3 @@
+[submodule "sqlite"]
+	path = sqlite
+	url = https://github.com/sqlite/sqlite