--- a/src/ffi.ts
+++ b/src/ffi.ts
@@ -430,7 +430,7 @@ try {
   if (customPath || Deno.build.os === "windows") {
     lib = Deno.dlopen(customPath || "sqlite3", symbols).symbols;
   } else {
-    const url = "https://github.com/denodrivers/sqlite3/releases/download/0.1.0/";
+    const url = "https://github.com/denodrivers/sqlite3/releases/download/v0.5.3/";
     lib = (await prepare({
       name: "sqlite3",
       urls: {