Towards https://github.com/denodrivers/sqlite3/issues/32

```
# main (m1)
Task bench deno run -A --unstable bench/bench_deno.js 50 1000000
time 155 ms rate 6451612
time 143 ms rate 6993006
time 142 ms rate 7042253
time 141 ms rate 7092198
time 141 ms rate 7092198
time 142 ms rate 7042253
time 142 ms rate 7042253
```

```
# This patch (m1)
Task bench deno run -A --unstable bench/bench_deno.js 50 1000000
time 137 ms rate 7299270
time 123 ms rate 8130081
time 122 ms rate 8196721
time 123 ms rate 8130081
time 122 ms rate 8196721
time 123 ms rate 8130081
time 122 ms rate 8196721
```